Add order, add order - supplier-product interaction

+++ app/controllers/productInfoCtrl.js
@@ -1,4 +1,4 @@
-app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams','productFactory',function($scope,$rootScope,$state,$stateParams,productFactory){
+app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams','productFactory','orderFactory','supplierFactory',function($scope,$rootScope,$state,$stateParams,productFactory,orderFactory,supplierFactory){
 
 	$scope.product;
 
@@ -6,6 +6,7 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 	$scope.stockAdjustAmount=0;
 	$scope.adjustReason='';
 	$scope.adjustDate='';
+	$scope.purchaseOrder=[];
 
 	var init= function(){
 		$scope.getItem();
@@ -20,6 +21,8 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 		});
 
 	}
+
+	$scope.init=init;
 	function changeWidth(){
 		var info=$('#sideinfo');
 		info.addClass('inInfo');
@@ -28,8 +31,9 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 		productFactory.getProduct($stateParams.productID)
 			.then(function(res){
 				$scope.product=res.data.data;
+				getOrder($scope.product.purchaseOrder);
 			},function(err){
-				window.warning('Cannot get product.\n Error message: '+ err.message);
+				alert('Cannot get product.\n Error message: '+ err.message);
 			});
 	};
 	$scope.editProduct= function(){
@@ -38,7 +42,7 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 				window.alert("Edit successfully.");
 				$scope.getItem();
 			},function(err){
-				window.warning("Cannot edit product.\n Error message: "+ err.message);
+				alert("Cannot edit product.\n Error message: "+ err.message);
 			});
 	}
 
@@ -54,24 +58,31 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 						window.alert("Adjust successfully.");
 						$scope.getItem();
 					},function(err){
-						window.warning("Cannot adjust product.\n Error message: "+ err.message);
+						alert("Cannot adjust product.\n Error message: "+ err.message);
 					});
 				
 			},function(err){
-				window.warning('Cannot get product.\n Error message: '+ err.message);
+				alert('Cannot get product.\n Error message: '+ err.message);
 			});
 	}
 
 	$scope.deleteItem=function(id){
 		var bool=window.confirm("The item will be deleted permanently. Do you want to proceed?");
+		if($scope.product.purchaseOrder.length>0)
+		{
+			alert("Please delete all related order before delete product.");
+			return;
+		}
 		if(bool==true){
 			productFactory.deleteProduct(id)
 				.then(function(res){
 					window.alert("Delete successfully.");
+					$state.go('product.list');
 				},function(err){
-					window.warning("Cannot delete product. \n Error message: "+err.message);
+					alert("Cannot delete product. \n Error message: "+err.message);
+					$state.go('product.list');
 				})
-			$state.go('product.list');
+			
 		}
 	}
 
@@ -79,6 +90,50 @@ app.controller("productInfoCtrl",['$scope','$rootScope','$state','$stateParams',
 		$scope.selectTab=tab;
 	}
 
+	function getOrder(orders){
+		$scope.purchaseOrder=[];
+		for(i in orders){
+			var tmp=orderFactory.getOrder(orders[i].orderID).then(function(res){
+				$scope.purchaseOrder.push(processOrder(res.data.data));
+			},function(err){
+				if(err){
+					alert("Cannot get orders of product!\n Error message: "+ err.message);
+				}
+			});
+		}
+	}
+
+	function processOrder(order){
+		var ret={};
+		ret._id=order._id;
+		ret.name=order.name;
+		ret.createdAt=order.createdAt;
+		ret.amount=0;
+		angular.forEach(order.batch,function(product){
+			if(product.productID==$scope.product._id){
+				ret.amount=product.amount;
+			}
+		});
+		if (order.isSolved==false){
+			ret.stat="Not Solved";
+		}
+		else {
+			ret.stat="Solved";
+		}
+
+		return ret;
+	}
+
+	function convertDate(string){
+		if (string==null)
+			return '';
+		var DateString=string.split('T')[0].split('-');
+		DateString.reverse();
+
+		return DateString.join('-');
+	}
+
+	$scope.convertDate=convertDate;
 	
-	init();
+	
 }])
